♻️ 컨트롤러 리팩토링, 사이드바 수정
