+++ src/main/java/com/takku/project/controller/SellerController.java
@@ -1,13 +1,16 @@
 package com.takku.project.controller;
 
-import com.takku.project.domain.CouponDTO;
+import com.takku.project.domain.StoreDTO;
 import com.takku.project.domain.UserDTO;
+import com.takku.project.domain.stats.OrderStatsDTO;
+import com.takku.project.domain.stats.PopularProductDTO;
+import com.takku.project.domain.stats.ProductRePurchaseDTO;
+import com.takku.project.domain.stats.TagStatsDTO;
+import com.takku.project.service.StoreService;
+import com.takku.project.service.StoreStatsService;
 import com.takku.project.service.UserService;
-
 import java.util.List;
-
 import javax.servlet.http.HttpSession;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,68 +22,109 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+
 @Controller
 @RequestMapping("/seller")
+@Api(tags = "판매자 관련 API")
 public class SellerController {
 
 	@Autowired
 	UserService userService;
 
-	// 1. 소상공인 내 정보 메인
+	@Autowired
+	private StoreService storeService;
+
+	@Autowired
+	private StoreStatsService statsService;
+
 	@GetMapping("/mypage")
+	@ApiOperation(value = "판매자 마이페이지", notes = "판매자의 기본 정보를 확인할 수 있는 마이페이지입니다.")
 	public String myPage(HttpSession session, Model model) {
 		UserDTO loginUser = (UserDTO) session.getAttribute("loginUser");
-
-		// 테스트용
-		UserDTO user = userService.selectByUserId(3);
-
+		// TODO 세션에서만 읽어오도록 수정
+		UserDTO user = userService.selectByUserId(3); // 테스트용
 		model.addAttribute("loginUser", user);
 		return "seller.mypage";
 	}
 
-	// 회원정보 수정 처리
 	@PostMapping("/mypage/update")
+	@ApiOperation(value = "판매자 정보 수정", notes = "판매자의 프로필 정보를 수정합니다.")
 	public String updateMyPage(@ModelAttribute UserDTO updatedUser, HttpSession session,
 			RedirectAttributes redirectAttributes) {
-		// 세션에서 현재 로그인한 사용자 정보 가져오기
 		UserDTO loginUser = (UserDTO) session.getAttribute("loginUser");
-
-		// 테스트용
-		UserDTO user = userService.selectByUserId(3);
-
+		// TODO 세션에서만 읽어오도록 수정
+		UserDTO user = userService.selectByUserId(3); // 테스트용
 		if (loginUser == null) {
-			// 로그인 세션 만료 시 로그인 페이지로 리다이렉트
 			return "redirect:/auth/login";
 		}
-
-		// 기존 로그인 사용자 ID 기준으로 값 보정
 		updatedUser.setUserId(user.getUserId());
-
-		// DB에 업데이트
 		userService.updateUser(updatedUser);
-
-		// 세션 정보도 최신값으로 갱신
 		session.setAttribute("loginUser", updatedUser);
-
-		// 알림 메시지 등 전달할 수 있음 (선택)
 		redirectAttributes.addFlashAttribute("updateSuccess", true);
-
 		return "redirect:/seller/mypage";
 	}
-	
+
 	@PostMapping("/partner/change")
 	@ResponseBody
+	@ApiOperation(value = "파트너 상태 변경", notes = "판매자의 파트너 등록/해제를 처리합니다.")
 	public String changePartnerStatus(@RequestParam("action") String action, HttpSession session) {
 		UserDTO loginUser = (UserDTO) session.getAttribute("loginUser");
-		// 테스트용
-		UserDTO user = userService.selectByUserId(3);
-		//if (loginUser == null) return "fail";
-
+		// TODO 세션에서만 읽어오도록 수정
+		UserDTO user = userService.selectByUserId(3); // 테스트용
 		String newStatus = action.equals("register") ? "Y" : "N";
 		user.setIsPartner(newStatus);
 		userService.updateUser(user);
 		session.setAttribute("loginUser", loginUser);
-
 		return "success";
+	}
+
+	@GetMapping("/home")
+	@ApiOperation(value = "판매자 홈 대시보드", notes = "판매자의 홈 화면에서 통계 데이터를 확인합니다.")
+	public String getMain(@RequestParam(required = false) String msg, Model model) {
+		// TODO 세션에서 읽어오도록 설정
+		Integer storeId = 1;
+		Integer userId = 1;
+		UserDTO user = userService.selectByUserId(userId);
+		StoreDTO store = storeService.selectStoreById(storeId);
+		int todayOrderCount = statsService.countTodayOrdersByStoreId(storeId);
+		int todaySales = statsService.sumTodaySalesByStoreId(storeId);
+		int ongoingFundingCount = statsService.countOngoingFundingsByStoreId(storeId);
+		int upcomingFundingCount = statsService.countUpcomingFundingsByStoreId(storeId);
+		model.addAttribute("userDTO", user);
+		model.addAttribute("storeDTO", store);
+		model.addAttribute("todayOrderCount", todayOrderCount);
+		model.addAttribute("todaySales", todaySales);
+		model.addAttribute("ongoingFundingCount", ongoingFundingCount);
+		model.addAttribute("upcomingFundingCount", upcomingFundingCount);
+		model.addAttribute("orderStats", statsService.getMonthlyOrderStats(storeId));
+		model.addAttribute("popularProducts", statsService.getPopularProducts(storeId));
+		model.addAttribute("tagStats", statsService.getTagStats(storeId));
+		model.addAttribute("topRePurchased", statsService.getTopRePurchasedProducts(storeId));
+		model.addAttribute("ageDistribution", statsService.getAgeDistribution());
+		model.addAttribute("genderRatio", statsService.getGenderRatio());
+		model.addAttribute("topTagsByGroup", statsService.getTopTagsByAgeGender());
+		return "seller.home";
+	}
+
+	@GetMapping("/stats")
+	@ApiOperation(value = "매장 통계 조회", notes = "판매자가 자신의 매장에 대한 통계를 조회합니다.")
+	public String getStoreStats(@RequestParam("storeId") int storeId, Model model, HttpSession session) {
+		// TODO 세션에서 읽어오도록 설정
+		Integer userId = 1;
+		UserDTO user = userService.selectByUserId(userId);
+		StoreDTO store = storeService.selectStoreById(storeId);
+		List<OrderStatsDTO> orderStats = statsService.getMonthlyOrderStats(storeId);
+		List<PopularProductDTO> popularProducts = statsService.getPopularProducts(storeId);
+		List<TagStatsDTO> tagStats = statsService.getTagStats(storeId);
+		List<ProductRePurchaseDTO> topRePurchased = statsService.getTopRePurchasedProducts(storeId);
+		model.addAttribute("userDTO", user);
+		model.addAttribute("storeDTO", store);
+		model.addAttribute("orderStats", orderStats);
+		model.addAttribute("popularProducts", popularProducts);
+		model.addAttribute("tagStats", tagStats);
+		model.addAttribute("topRePurchased", topRePurchased);
+		return "seller/stats";
 	}
 }
